Use test flags instead
